+++ breeze-reflect/src/main/kotlin/com/windea/breezeframework/reflect/extensions/CoreExtensions.kt
@@ -1,0 +1,41 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
+package com.windea.breezeframework.reflect.extensions
+
+import com.windea.breezeframework.core.annotations.api.*
+import kotlin.reflect.*
+
+/**得到指定类型的名字。*/
+@TrickImplementationApi("Can never be implemented for all situations.")
+@ExperimentalStdlibApi
+inline fun <reified T> nameOf(): String? {
+	return T::class.simpleName
+}
+
+/**得到指定项的名字。适用于：类型、属性引用、方法引用、实例。不适用于：类型参数，参数，局部变量。*/
+@TrickImplementationApi("Can never be implemented for all situations.")
+@ExperimentalStdlibApi
+inline fun nameOf(target: Any?): String? {
+	return when {
+		target == null -> null
+		target is KClass<*> -> target.simpleName
+		target is KCallable<*> -> target.name
+		//无法直接通过方法的引用得到参数
+		target is KParameter -> target.name
+		//无法得到局部变量的任何信息
+		else -> target::class.simpleName
+	}
+}
+
+
+/**得到指定类型的引用。等同于`this::class`。*/
+@TrickImplementationApi
+inline fun <reified T> refOf() {
+	TODO()
+}
+
+/**得到指定项的引用。参见引用操作符`::`。*/
+@TrickImplementationApi
+inline fun refOf(target: Any?) {
+	TODO()
+}